+++ packages/client/src/common/types.d.ts
@@ -13,14 +13,15 @@ export type ShortcutDataProps = {
 	url: string;
 };
 
-export type ShortcutProps = {
-	data: ShortcutDataProps[];
+export type SectionProps = {
+	id: string;
 	position: number;
+	shortcuts: ShortcutDataProps[];
 	title: string;
 };
 
 export type StateProps = {
-	shortcuts: ShortcutProps[];
+	sections: SectionProps[];
 	status:
 		| string
 		| typeof ACTION_STATUS_STALE
@@ -41,7 +42,7 @@ export type ActionProps =
 	  }
 	| {
 			payload: {
-				shortcuts: ShortcutProps[];
+				sections: SectionProps[];
 				status:
 					| string
 					| typeof ACTION_STATUS_ERROR
@@ -51,7 +52,7 @@ export type ActionProps =
 	  }
 	| {
 			payload: {
-				shortcut: ShortcutProps;
+				section: SectionProps;
 				status: typeof ACTION_STATUS_STALE;
 			};
 			type: typeof ACTION_SET_DATA;